fix: 串行的Stage审核超时的流水线在多次重试后会导致流水线卡住 #5090 防溢出

+++ src/backend/ci/core/process/api-process/src/main/kotlin/com/tencent/devops/process/engine/common/Timeout.kt
@@ -39,11 +39,11 @@ object Timeout {
 
     val MAX_HOURS = TimeUnit.DAYS.toHours(MAX_STAGE_REVIEW_DAYS) // 60 * 24 = 1440 小时 = 审核最多超时60天
 
-    val STAGE_MAX_MILLS = TimeUnit.HOURS.toMillis(MAX_HOURS).toInt() + 1 // 毫秒+1
+    val STAGE_MAX_MILLS = TimeUnit.HOURS.toMillis(MAX_HOURS) + 1 // 毫秒+1
 
     val MAX_MINUTES = TimeUnit.DAYS.toMinutes(MAX_JOB_RUN_DAYS).toInt() // 7 * 24 * 60 = 10080 分钟 = 最多超时7天
 
-    val CONTAINER_MAX_MILLS = TimeUnit.MINUTES.toMillis(MAX_MINUTES.toLong()).toInt() + 1 // 毫秒+1
+    val CONTAINER_MAX_MILLS = TimeUnit.MINUTES.toMillis(MAX_MINUTES.toLong()) + 1 // 毫秒+1
 
     fun transMinuteTimeoutToMills(timeoutMinutes: Int?): Pair<Int, Long> {
         var minute = timeoutMinutes ?: DEFAULT_TIMEOUT_MIN
